% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgbysample.R
\name{avgbysample}
\alias{avgbysample}
\title{Compute Avg. or Mean by Sample Size. Takes data column and samplesize as input and computes the avg or mean by samplesize as output dataframe.}
\usage{
avgbysample(data, samplesize)
}
\arguments{
\item{data}{numeric}

\item{samplesize}{numeric}
}
\value{
dataframe of avg or mean values by samplesize.
}
\description{
Compute Avg. or Mean by Sample Size. Takes data column and samplesize as input and computes the avg or mean by samplesize as output dataframe.
}
\examples{
avgbysample(c(1,3,5,9),2)
avgbysample(c(1,3,5,9,10,15,19,18),4)
avgbysample(iris$Sepal.Length,5)
}
